Allow for files that get updated as time goes by
Previously, we'd been assuming that once files appear, they no longer
are updated. This is probably wrong.

Instead, check if this file is the most recent one for a given room. If
so, that means the file may have changed size, so check that, and then
update the metadata in the database if things have indeed changed.