Don't base the end time on the parsed time in case of update
We're currently testing things by faking timestamps. The detect_files
script, however, sees that (somewhat) and then overwrites the endtime
field, but not the starttime field, breaking our tests in the process.

Fix by having the database calculate the endtime based on the length of
the video and the starttime in the database, rather than the starttime
as passed by detect_files.

As a side effect, this also allows us to override autodetected start
times, which may come in handy at some point in the future.