Make less assumptions
We were currently assuming that the "pre" or "post" videos would
*always* be exactly 20 minutes long. This might not happen, e.g., if the
talk was the first of the day and the video recording only started five
minutes before.

Rather than assuming 20 minutes, figure out what the actual length of
the video is at cut time, and base our calculations on that.