Add license notices
Just having a COPYRIGHT file is not usually enough, so add explicit
license notices to every file that contains code.

While at it, make the license notice in the HTML file be
LibreJS-compliant, so that people who use that will be happy, too. This
will be used by FOSDEM, after all.