Create a module for supporting configuration files
This module will only return configuration variables that were
previously defined, and will allow outputting a template configuration
file based on those definitions.

When requesting a configuration value, values that are not defined in
the configuration file are returned as their default value rather than
the configured value. At the same time, we do not complain if a
configuration value is defined that we do not know about.

This allows us to easily self-document scripts, and to merge
configuration values for multiple scripts into one file.