Update the old-style debconf script to (hopefully) work
Yes, we already wrote backend/cut_talk which is supposed to replace
both. However, that wants the new config interface, and rewriting
*everything* for the new config interface is going to take a lot of work
and not be ready in time.

So for now, keep using the old scripts, but fold the improvements we
made for FOSDEM back into the (older) debconf scripts.

Start with cut_talk, do the rest in the next commits.