Add empty audio track to slide videos
ffmpeg documentation states that concatenated files need to have the
same number of tracks; however, it doesn't produce an error message if
that is not followed, and silently fixes up the resulting video. When
the first two videos differ in number of audio tracks, apparently ffmpeg
adds an empty audio track to the beginning of the video and moves on.

When they do not, however, no audio track is added to the video,
resulting in a video file with no audio. We trigger this when we add an
apology note, since then we have two files with no audio as the first
two files.

Fix by explicitly adding an (empty) audio track to all video files.