Allow for a (released) version number
Currently we assume that everything runs from git. While this is true
for the time being, it won't be once we package this and distribute it
further. As such, allow for "released" version numbers being used.

Drop the "use DBI" in the SReview main package, it's of no use there.

Modify Makefile.PL so it picks the version number from SReview.pm
instead of hardcoding it in there.