Fix outputdir assignment
Originally, outputdir was an anonymous sub that could do whatever it wanted.

In this particular case, subs are not a good idea for various reasons, however:
- They do not work with Data::Dumper (so would get lost on every update)
- They are not really necessary
- They are confusing for people who do not understand Perl very well.

As such, change this to a static list of subdirectory elements that the
user can select; that should allow for enough flexibility, and can
always be revisited if necessary.