Add a profile for FOSDEM
Do so not in the main profile factory class, since we may not stick with
it -- and we don't really need to use it everywhere.

Update SReview::Video::ProfileFactory so it can load classes when
needed.

Also, extend the test so that we check that the FOSDEM class can be used
(so that extensibility is tested)