Fix test for clobbering database, and reorder queries
We tried to clobber the database by using Mojo::Pg migrations, but did
it wrong, and the database wasn't being clobbered. When we fixed that,
it turned out that the order of some of the queries in the "4 down"
migration were broken, in that we tried to drop objects that were still
being used by other objects that we tried to delete a few lines later.

Reorder things so that they will work, and fix the database check so
that this case is being tested for.