Detect whether libfdk_aac exists
If ffmpeg has been built with libfdk_aac enabled, then it will tell you
that the aac codec is experimental and that you should use that one
please, unless you explicitly tell it to use libfdk_aac.

We should at some point chage SReview::Video so you can explicitly tell
ffmpeg not to use libfdk_aac; but until then, detect at run time whether
it was compiled in, and add the required bits to the codec map if not.

While this may not be a bad idea to do in general, we should probably
make this be a singleton Moose object that only calls the necessary
ffmpeg things when we're actually trying to read them. That's for later,
though.