Add metadata to generated transcoded files
While at it, also rename "startdate" to "date", since we expect that an
event starts and ends on the same date (if not... well, too bad).

Might want to update @STARTTIME@ too at some point, but not now.

That has a minor interface difference, but we're 0.x still anyway, so
whatever.