Change defaults
Currently, SReview defaults to using gridengine as an external
scheduler.

While that is a good idea in general, the problem with that approach is
that it requires us to set up gridengine, even if it's just for a small
event. This may be a bit of a roadblock for people unaccustomed to such
an external scheduler.

Instead, still recommend to use it, but default to not doing so.