Add the possibility to create profiles from config
Currently, profiles can only be created by subclassing
SReview::Video::Profile::Base (or one of its children). We don't want to
make that impossible, but for the common and easy case, that does seem
like a lot of busywork.

Instead, make it so that we can create one from the configuration file
by a simple declarative syntax, and also test from the test suite that
it actually works.

Signed-off-by: Wouter Verhelst <w@uter.be>