Improve database initialization
Currently, whenever we want to change a field in a table that is also
used by one of the stored procedures, we have to drop the stored
procedure in the migration, alter the table, and then recreate the
stored procedure. This is mildly annoying, and creates a lot of useless
repetition of functions in the migrations.

Instead, separate the "code" and "data" migrations from one another, so
that we can do

- drop all stored procedures
- update the data migrations to the latest version
- upgrade the stored procedures to the latest version

This avoids the need to repeat whole stored procedures when we want to
touch some tables.