Introduce the stream concept
Sometimes something bad happens to a talk's recording. For instance, the
original recordings may be lost, and we may need to fall back to a
lower-quality version of the same talk, or to an alternate recording
which may be incomplete, and/or whose timestamps may be off from the
original one.

For instance, we may have the following options:

- Original high-quality/high bandwidth archival quality recordings from
  voctomix
- A dump of the stream as it was sent over the Internet, including
  interruptions like sponsor loops etc (which we would prefer not to
  release over the Internet)
- A not-mixed plain dump of the camera output, which we may need to fall
  back to in some cases.

Currently, whenever we need to switch to an alternate stream, we make a
"fake" room which contains the alternate stream, and then move the talk
in question to the fake room. This works, but means we either need to
write the schedule parser in such a way that it would ignore those fake
rooms, or we need to make sure that the schedule parser won't run
anymore after the room in question has been created -- otherwise things
go horribly wrong. By introducing a stream concept, we can switch a talk
to an alternate stream, with no issues for that.

TODO:
- Add configuration options so that input profiles may differ depending
  on which stream is in use;
- Add UI to switch a room to an alternate stream from the webinterface.