Add a "serial" property
It is useful to be able to track how many changes have been made to a
given talk, even if we don't keep track of the individual changes. Do so
by adding a "serial" property.

The alternative is to have a serial number for every change, and then
use max() in the query that tries to get the most recent correction
values, but that will blow up the size of that table, which seems like a
bad idea, performance-wise, for very little benefit.

TODO: migrate this to an enum. There is little benefit to do it like
this anymore, now that we're not dynamically showing the properties
anymore...