Fix queries in face of no corrections yet
When the corrections table was empty for a given talk, these queries did
not return anything, resulting in the JSON data that we send to the user
missing items.

Fix up the joins, and add use of the "coalesce" function where
necessary, to remedy.

While at it, drop talks.id from the select list, which we don't end up
using.