Reconnect to the database after large amounts of work
Currently we use one connection for the whole of the trascode script.

This may occasionally not work: the database server may have restarted,
or there may be a bad router in between which snacks on TCP connections
for lunch.

Assume that after doing something that may take several hours, our
connection is no longer valid.