Move the property filter to the WHERE clause
Apparently the optimizer may throw out joined tables if they're not used
after the FROM clause.

We don't really care about the data from the "properties" table, except
that we want to filter on it. Since that means the table isn't used
after the FROM clause, it means we get far too much data.

Move it to the WHERE clause, so it doesn't get thrown out and so we only
see the data that we care about.