sreview-previews: don't skip *everything*
SReview::Files::Collection::has_file expects one (positional) argument,
rather than a hash.

Forgetting that means that the positional argument becomes the hash key,
which means we try to look for a file named "relname" or some such.

Fix.

While at it, also fix the test suite so it will fail in a case like
this. We need to look for the output of sreview-previews, not its
input...