sreview-import: implement
This is a generic schedule parser, which separates "parse the upstream
schedule" from "make sure the database is up to date". Add two parsers;
one for Penta XML, and one for the Wafer version of the same, which
derives from the Penta parser (but changes the bits that Wafer breaks).

Currently sreview-import moves all events that disappear from the
upstream schedule to the "ignored" state if they're still in
"waiting_for_files". Perhaps we can change that later, if this is deemed
necessary.