sreview-cut: rewrite.
This is an experimental rewrite of sreview-cut, organized in a better
way, with some of the lower-level logic moved to separate modules.

Uses the SReview::Normalizer (and SReview::Normalizer::Bs1770gain)
modules for normalizing audio, and a (not yet complete) module for
fixing A/V desync.

Currently, new implementation (seems to) work for everything, except A/V
desync.