Add an FFmpeg "loudnorm" based audio normalizer
bs1770gain works, but has Issues(TM) that we would prefer not to have to
deal with.

Implement an audio normalizer that uses ffmpeg's "loudnorm" filter,
rather than bs1770gain.

Keep bs1770gain as the default normalizer for now.

There's also a TODO in the ffmpeg filter, but let's make sure we have
something for now, first; we can refactor things later if we need to.