Make autopkgtest not fail over templates
Test::Mojo doesn't run a Mojolicious app in production mode, which means
that the files under /usr/share won't be served.

That's fine when we run the test suite from the build directory, but it
makes things fail when we try to run it from autopkgtest.

To fix, add the directories if they appear to exist rather than based on
the mode we're running things in.