Account for variation in ffprobe output
Sometimes ffprobe outputs the video bitrate in bits per second; other
times it outputs it in kilobits per second. It is not entirely clear to
us how this is decided; it might be a version thing, or it might be
different depending on details of the file being read (container, codec,
...)

SReview::Video assumes kilobits internally; when a video is examined
that returns a video bitrate in bits per second rather than kilobits per
second, we may end up issuing an ffmpeg command line that sets a video
bitrate that is off by 3 orders of magnitude. If the bitrate was high to
begin with, this may make ffmpeg return an overflow error, resulting in
an empty file.

Fix by (conditionally) dividing the returned bit rate.