Publication status detection: allow for multiple header values
Previously, a video file would be marked as "published" if all given
headers and/or the HTTP result code had a particular given value.

In some cases, however, there are multiple values that would indicate
that the video is available.

Allow for specifying multiple values to allow for that case.