Stop creating file object with absolute filenames
We originally didn't have the SReview::Files API, and everything went
through files on the filesystem, using NFS if we needed to access hosts
on the network, which required that we mounted filesystems in the exact
same location.

The SReview::Files API makes it possible to remap things (e.g., from
filesystem to HTTP URLs), provided we don't use absolute URLs but
relative ones.