SReview::Files::Collection::HTTP::nginx::JSON: add.
nginx has a JSON index format that provides computer-readable metadata
on files in its export. Create a subclass of
SReview::Files::Collection::HTTP that implements _probe_children using
that metadata.

Can only be used for an input collection, obviously, but at least this
allows us to use it for sreview-detect purposes.