Unbreak the test suite
At FOSDEM, we're currently using multiple audio streams. This results in
ffmpeg being called with "-map 0:v -map 0:<audio channel>", and that
works.

However, for stereo encoding of the audio channels, this doesn't work,
as then we don't have an explicit audio map, so we get a video file with
no audio, instead of what we want.

Make changes so that audio extraction is done in one step, and then the
next we merge with the video file again. Not ideal, but we have no
better option.