Add two new states
"fixuping": to allow running scripts to fix up the videos after
transcoding, without requiring a full transcode run.

Can be used for:

- Muting audio on the opening/closing slides, if necessary
- Muxing in subtitles
- Syncing metadata
- And so much more.

Yes, the name is juvenile. Sue me.

"uninteresting": alternate name for "ignored", but designed so that it
is not used by sreview-import and considered as "oh look it's back on
the schedule, let's move it to waiting_for_files".