Refactor creation of objects
It's easier to subclass a parser if the helper classes (such as ones for
speakers, talks, and rooms) are configurable, too.

We had already done some of this in some of our subclasses (e.g.,
WithShadow and Penta), but not consitently so.

Move the configurability over to the base class, and rewrite every
"SReview::Schedule::Foo::Bar->new" into "$self->foo_type->new", so that
we can configure the exact class from a child class (or from SReview
configuration, as the case may be).