Refactor again
Copying a reference to each child object is a lot of work, and requires
some messiness.

Instead, give every object in the tree a (weak) reference to its parent,
so that we can fetch data pertaining to ourselves from the parent, if
required.

This also makes it easier to subclass; you don't need to subclass the
Talk or Event objects if you don't need to modify them, and can just
fetch data from $self->talk_object and/or
$self->talk_object->event_object and/or
$self->talk_object->event_object->root_object, as required.