Make canonical duration element configurable
Unfortunately some versions of ffprobe misdetect the duration of a
video. Unfortunately some misdetect the container duration while some
others misdetect the video content duration, so we can't rely on one or
the other being correct always :-(

At least if it misdetects things, it is consistently incorrect in the
same way everywhere, so we can default to the element that makes the
most sense (the container) while allowing particular instances to use a
different setting.