Fix error handling in OpenAPI code
The OpenAPI plugin validates input based on the "status" bit, but we've
been setting ->res->code everywhere, which it does not look at.

The result is that the validator fails on the objects when we pass them,
since it expects the 200 body rather than the error body.

We've been bypassing the OpenAPI validator by rendering plain text, but
that confuses client libraries.

Do it right.