Drop the /released full replacement
We have a /released endpoint. It's still there.

We were working on a replacement for it, under the structured API. That
doesn't work, and it's too complicated and not very REST-y. Get rid of
it.

In place of the call that was trying to get you everything, add a call
that allows you to get a list of known talks in an event that have a
given state.

We will add the rest of the needed calls in future work