Improve warning message for when files don't exist
The warning message which we issued when a file did not exist was
confusing; people who did not write this code often mistook it for an
error that could cause a problem, when it was merely being issued when
we deleted files just before writing the new ones.

Expand the message so it is clear that it is a file deletion request and
that it is being ignored because the file is already not there.