Don't create empty files
When we try to create a new key frame for a cut point, we would only
avoid doing that if the cut point position is *exactly* on a key frame.

Due to how things work, sometimes we may have a request to create a cut
point at a point in time in between frames. If this happens to be less
than a frame before a key frame, the result will then be a zero-sized
"recoded" file, which ffmpeg barfs on when we try to concatenate the
files later.

Fix by exiting early if the requested cut point is less than a frame
before a key frame.