Account for rounding errors
The desired cut point is not guaranteed to be exactly on a frame. If we
ask for a forced key frame, and the desired cut point is closer to a
frame that just passed rather than the frame that is coming, ffmpeg will
round down.

This will cause us to request ffmpeg to cut just after that
newly-created key frame, which it will interpret as a cut from after
that key frame rather than from the key frame, which is exactly what we
don't want.

Fix that by searching for key frames in the stream again, comparing them
with our desired cut point, and doing a cut from the key frame that is
the closest to our cut point (in absolute value).