SReview::Files::Factory: start documenting this better
The SReview::Files API is simple, but not easy to understand if you
don't know how it works.

Add documentation to make it easier to understand.

This does require that the classes are split up properly, otherwise the
POD stuff ends up being all smushed together.

TODO: finish this. Soon, I hope.